Implement F16 : F20 && F23 And edit System data member access

+++ StringFunctions.h
@@ -7,8 +7,8 @@ class StringFunctions {
 public:
 	static vector<string> split(string str, char delimiter) {
 		vector<string> itemList;
-		int start = 0;
-		int end = str.find(delimiter);
+		size_t start = 0;
+		size_t end = str.find(delimiter);
 		while (end <= str.length()) {
 			itemList.push_back(str.substr(start, end - start));
 			start = end + 1;
